Split QP/Conic tests
